cleaned up project and added some more handler methods

+++ includes/tcp.h
@@ -19,15 +19,15 @@ enum tcp_flag{
 typedef struct TCPPack{
     uint8_t source_port;
     uint8_t destination_port;
-    uint32_t sequence_number;
-    uint32_t acknowledgement_number;
+    uint8_t sequence_number;
+    uint8_t acknowledgement_number;
     enum tcp_flag flag;
-    uint16_t advertisement_window;
+    uint8_t advertisement_window;
     uint32_t checksum;
-    uint8_t payload[MAX_PAYLOAD_SIZE];
+    uint16_t payload[MAX_PAYLOAD_SIZE];
 }TCPPack;
 
-void makeTCPPacket(TCPPack* TCP, uint16_t source_port, uint16_t destination_port, uint32_t sequence_number, uint32_t acknowledgement_number, uint16_t flag, uint16_t advertisement_window, uint32_t checksum, uint8_t* payload) {
+void makeTCPPacket(TCPPack* TCP, uint8_t source_port, uint8_t destination_port, uint8_t sequence_number, uint8_t acknowledgement_number, uint8_t flag, uint8_t advertisement_window, uint32_t checksum, uint8_t* payload) {
     TCP->source_port = source_port;
     TCP->destination_port = destination_port;
     TCP->sequence_number = sequence_number;
